+++ src/main/scala/com/lucendar/common/db/rest/PaginationResolver.scala
@@ -7,9 +7,9 @@
  * ******************************************************************************/
 package com.lucendar.common.db.rest
 
+import com.lucendar.common.types.rest.Pagination
+import com.lucendar.common.utils.StringUtils
 import info.gratour.common.error.ErrorWithCode
-import info.gratour.common.types.rest.Pagination
-import info.gratour.common.utils.StringUtils
 import org.springframework.core.MethodParameter
 import org.springframework.web.bind.support.WebDataBinderFactory
 import org.springframework.web.context.request.NativeWebRequest
@@ -47,7 +47,7 @@ class PaginationResolver(maxLimit: Int) extends HandlerMethodArgumentResolver {
     })
 
     if (limit != null && page != null) {
-      Pagination(limit, page)
+      new Pagination(limit, page)
     } else
       null
   }