+++ src/main/scala/com/lucendar/common/db/types/JavaByteArray.java
@@ -1,0 +1,13 @@
+/** *****************************************************************************
+ * Copyright (c) 2019, 2021 lucendar.com.
+ * All rights reserved.
+ *
+ * Contributors:
+ * KwanKin Yau (alphax@vip.163.com) - initial API and implementation
+ * ******************************************************************************/
+package com.lucendar.common.db.types;
+
+public class JavaByteArray {
+
+    public static Class<?> BYTE_ARRAY_CLASS = byte[].class;
+}